# starter/gin/docs/swagger.yaml
basePath: /api
definitions:
  audit.AuditActionDTO:
    properties:
      action:
        description: '"user.create"'
        type: string
      category:
        description: '"user"'
        type: string
      description:
        description: '"Create new user"'
        type: string
      label:
        description: '"创建用户"'
        type: string
      operation:
        description: '"create"'
        type: string
    type: object
  audit.AuditActionsResponseDTO:
    properties:
      actions:
        description: 操作定义列表
        items:
          $ref: "#/definitions/audit.AuditActionDTO"
        type: array
      categories:
        description: 分类选项
        items:
          $ref: "#/definitions/audit.CategoryOptionDTO"
        type: array
      operations:
        description: 操作类型选项
        items:
          $ref: "#/definitions/audit.OperationTypeDTO"
        type: array
    type: object
  audit.AuditDTO:
    properties:
      action:
        type: string
      created_at:
        type: string
      details:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      resource:
        type: string
      status:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
  audit.CategoryOptionDTO:
    properties:
      label:
        description: '"用户管理"'
        type: string
      value:
        description: '"user"'
        type: string
    type: object
  audit.OperationTypeDTO:
    properties:
      label:
        description: '"创建"'
        type: string
      value:
        description: '"create"'
        type: string
    type: object
  auth.Login2FADTO:
    properties:
      session_token:
        description: 登录时返回的临时会话令牌
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      two_factor_code:
        description: 6位TOTP验证码
        example: "123456"
        type: string
    required:
      - session_token
      - two_factor_code
    type: object
  auth.LoginDTO:
    properties:
      account:
        description: 手机号/用户名/邮箱
        example: admin
        type: string
      captcha:
        description: 验证码
        example: "9999"
        type: string
      captcha_id:
        description: 验证码ID
        example: dev-123456
        type: string
      password:
        description: 密码
        example: admin123
        type: string
    required:
      - account
      - captcha
      - captcha_id
      - password
    type: object
  auth.LoginResponseDTO:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      requires_2fa:
        description: 2FA 相关（当需要 2FA 时返回）
        type: boolean
      session_token:
        type: string
      token_type:
        type: string
      user:
        $ref: "#/definitions/auth.UserBriefDTO"
    type: object
  auth.LoginRoleDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  auth.RefreshTokenDTO:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
      - refresh_token
    type: object
  auth.RegisterDTO:
    properties:
      email:
        example: john@example.com
        type: string
      nickname:
        example: Johnny
        maxLength: 50
        type: string
      password:
        example: password123
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        type: string
      real_name:
        example: John Doe
        maxLength: 100
        type: string
      signature:
        example: Hello World
        maxLength: 255
        type: string
      username:
        example: john_doe
        maxLength: 50
        minLength: 3
        type: string
    required:
      - email
      - password
      - username
    type: object
  auth.RegisterResultDTO:
    properties:
      access_token:
        type: string
      email:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  auth.TokenDTO:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  auth.UserBriefDTO:
    properties:
      roles:
        items:
          $ref: "#/definitions/auth.LoginRoleDTO"
        type: array
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  response.DataResponse-audit_AuditActionsResponseDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
          - $ref: "#/definitions/audit.AuditActionsResponseDTO"
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-audit_AuditDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
          - $ref: "#/definitions/audit.AuditDTO"
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-auth_LoginResponseDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
          - $ref: "#/definitions/auth.LoginResponseDTO"
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-auth_RegisterResultDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
          - $ref: "#/definitions/auth.RegisterResultDTO"
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-auth_TokenDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
          - $ref: "#/definitions/auth.TokenDTO"
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: "#/definitions/response.ErrorDetail"
    type: object
...
  response.PagedResponse-audit_AuditDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 数据列表
        items:
          $ref: "#/definitions/audit.AuditDTO"
        type: array
      message:
        description: 消息描述
        type: string
      meta:
        allOf:
          - $ref: "#/definitions/response.PaginationMeta"
        description: 分页信息
    type: object
...
  response.PaginationMeta:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      per_page:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
      warning:
        description: 页码越界警告
        type: string
    type: object
...
host: localhost:40012
info:
  contact:
    email: lwmacct@icloud.com
    name: API Support
    url: https://github.com/lwmacct/260101-go-pkg-ddd/issues
  description: 基于 DDD+CQRS 架构的 Go Web 应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lwmacct/260101-go-pkg-ddd
  title: Go DDD Template API
  version: 1.0.3
paths:
  /api/admin/audit:
    get:
      consumes:
        - application/json
      description: 分页获取审计日志，支持按用户、操作、资源、状态、时间范围筛选
      parameters:
        - description: Action 操作类型过滤（语义化标识，如 setting.update）
          in: query
          name: action
          type: string
        - description: EndDate 结束时间（RFC3339 格式）
          in: query
          name: end_date
          type: string
        - default: 20
          description: Limit 每页数量，默认 20，最大 1000
          in: query
          maximum: 1000
          minimum: 1
          name: limit
          type: integer
        - default: 1
          description: Page 页码，从 1 开始
          in: query
          minimum: 1
          name: page
          type: integer
        - description: Resource 资源分类过滤（如 setting, user）
          in: query
          name: resource
          type: string
        - description: StartDate 开始时间（RFC3339 格式）
          in: query
          name: start_date
          type: string
        - description: Status 状态过滤
          enum:
            - success
            - failure
          in: query
          name: status
          type: string
        - description: UserID 按用户 ID 过滤
          in: query
          name: user_id
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 审计日志列表
          schema:
            $ref: "#/definitions/response.PagedResponse-audit_AuditDTO"
        "400":
          description: 参数错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 审计日志列表
      tags:
        - Admin - Audit Log
  /api/admin/audit/{id}:
    get:
      consumes:
        - application/json
      description: 根据日志ID获取审计日志详细信息
      parameters:
        - description: 日志ID
          in: path
          minimum: 1
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 日志详情
          schema:
            $ref: "#/definitions/response.DataResponse-audit_AuditDTO"
        "400":
          description: 无效的日志ID
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "404":
          description: 日志不存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 审计日志详情
      tags:
        - Admin - Audit Log
  /api/admin/audit/actions:
    get:
      consumes:
        - application/json
      description: 获取所有审计操作的定义、分类和操作类型，供前端筛选器使用
      produces:
        - application/json
      responses:
        "200":
          description: 审计操作定义
          schema:
            $ref: "#/definitions/response.DataResponse-audit_AuditActionsResponseDTO"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 审计操作定义
      tags:
        - Admin - Audit Log
...
  /api/auth/login:
    post:
      consumes:
        - application/json
      description: 使用手机号/用户名/邮箱和密码登录系统，需要提供图形验证码。如果启用了2FA，返回session_token用于后续2FA验证
      parameters:
        - description: 登录凭证
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/auth.LoginDTO"
      produces:
        - application/json
      responses:
        "200":
          description: 登录成功或需要2FA验证
          schema:
            $ref: "#/definitions/response.DataResponse-auth_LoginResponseDTO"
        "401":
          description: 登录失败：凭证无效、验证码错误或账户被禁用
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      summary: 登录
      tags:
        - Authentication
  /api/auth/login/2fa:
    post:
      consumes:
        - application/json
      description: 使用session_token和2FA验证码完成登录（适用于启用了2FA的账户）
      parameters:
        - description: 二次认证凭证
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/auth.Login2FADTO"
      produces:
        - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: "#/definitions/response.DataResponse-auth_TokenDTO"
        "401":
          description: 验证失败：session_token无效或2FA验证码错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      summary: 2FA 登录
      tags:
        - Authentication
  /api/auth/refresh:
    post:
      consumes:
        - application/json
      description: 使用refresh_token获取新的access_token和refresh_token，延长会话有效期
      parameters:
        - description: 刷新令牌
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/auth.RefreshTokenDTO"
      produces:
        - application/json
      responses:
        "200":
          description: 令牌刷新成功
          schema:
            $ref: "#/definitions/response.DataResponse-auth_LoginResponseDTO"
        "401":
          description: 刷新令牌无效或已过期
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 刷新令牌
      tags:
        - Authentication
  /api/auth/register:
    post:
      consumes:
        - application/json
      description: 创建新用户账号，注册成功后自动登录并返回访问令牌
      parameters:
        - description: 注册信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/auth.RegisterDTO"
      produces:
        - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: "#/definitions/response.DataResponse-auth_RegisterResultDTO"
        "400":
          description: 参数错误或用户名/邮箱已存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      summary: 注册
      tags:
        - Authentication
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"